backend/requirements.txt
yfinance
scikit-learn
pandas
fastapi
uvicorn
google-genai